k8s/db_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
prometheus-client